// src/app/models/unit.model.ts
import { Restaurant } from "./restaurant.model";
import { Size } from "./size.model";
import { User } from "./user.model";

export interface Unit {
    id: number;
    name: string;
    description: string;
    created: Date;
    updated: Date;
    deleted: boolean;
    restaurant: Restaurant;
    createdUser: User;
    updatedUser: User;
}